Added arbitrary precision signed decimal numbers
Added arbitrary precision signed decimal numbers

+++ src/main/java/com/challenge/calculator/server/Server.java
@@ -1,5 +1,7 @@
 package com.challenge.calculator.server;
 
+import java.math.BigDecimal;
+
 import org.springframework.amqp.rabbit.annotation.RabbitListener;
 
 import com.challenge.calculator.enums.OperationsEnum;
@@ -8,8 +10,8 @@ import com.challenge.calculator.model.OperationModel;
 public class Server {
 
 	@RabbitListener(queues = "calculator.rpc.requests")
-	public Double sum(OperationModel model) {
-		Double result = null;
+	public BigDecimal sum(OperationModel model) {
+		BigDecimal result = null;
 		System.out.println(" [x] Received request for ");
 		OperationsEnum op = model.getOperation();
 		result = op.calculate(model.getA(), model.getB());
